Allow writing of node usages for reach fp
Writing node usages for Reach FP was not the issue here. Instead it was the order of bones in the granny file. Now having set the latter correctly the node usages can now be used. I assume without this set the game gets the bones to use for poses by name